Include an error case
The most common error is going to be "BadRequest" due to a bad address,
this happens when the string address passed to the request is not a
valid address string.

We don't test for success because that would require a valid address
string and we don't want to inundate the network with garbage/test
accounts, nor do we want to expose a valid address to the public.